mdmon: fix missed 'clean' event
mdmon may miss events because it re-reads state after read_and_act.  The
additional read is used to determine dirty status before allowing a
sigterm to proceed.  Since read_and_act is in the best position to
determine 'dirty' status and its return value is not used, modify it to
return true if the array is dirty.

Signed-off-by: Dan Williams <dan.j.williams@intel.com>